Uniformisation de WebRSA
- Ajout de la structure complète de la base de données
- Ajout des données minimales pour la base de données
- Suppression du cast (int) pour le département
- Ajout de la prise en compte des départements contenant une lettre

Issue: #80137

+++ project/app/Controller/ContratsinsertionController.php
@@ -410,7 +410,7 @@
 		 * @param integer $personne_id L'id technique de la personne.
 		 */
 		public function index($personne_id = null) {
-			$departement = (int)Configure::read('Cg.departement');
+			$departement = Configure::read('Cg.departement');
 			$dossierMenu = $this->DossiersMenus->getAndCheckDossierMenu(array('personne_id' => $personne_id));
 			$this->_setEntriesAncienDossier($personne_id, 'Contratinsertion');
