+++ project/app/Utility/WebrsaAccessContratsinsertion.php
@@ -19,42 +19,42 @@
 	{
 		/**
 		 * Paramètres par défaut
-		 * 
+		 *
 		 * @param array $params
 		 * @return array
 		 */
 		public static function params(array $params = array()) {
 			return $params + array(
 				'alias' => 'Contratinsertion',
-				'departement' => (int)Configure::read('Cg.departement')
+				'departement' => Configure::read('Cg.departement')
 			);
 		}
 
 		/**
 		 * Permission d'accès
-		 * 
+		 *
 		 * @param array $record
 		 * @param array $params
 		 * @return boolean
 		 */
 		protected static function _add(array $record, array $params) {
 			$params = self::params($params);
-			
+
 			switch ($params['departement']) {
 				case 58:
 				case 66:
 					$access = Hash::get($params, 'ajoutPossible');
 					break;
-				default: 
+				default:
 					$access = true;
 			}
-			
+
 			return $access;
 		}
 
 		/**
 		 * Permission d'accès
-		 * 
+		 *
 		 * @param array $record
 		 * @param array $params
 		 * @return boolean
@@ -65,55 +65,55 @@
 
 		/**
 		 * Permission d'accès
-		 * 
+		 *
 		 * @param array $record
 		 * @param array $params
 		 * @return boolean
 		 */
 		protected static function _edit(array $record, array $params) {
 			$params = self::params($params);
-			
+
 			switch ($params['departement']) {
-				case 66: 
+				case 66:
 					$access = Hash::get($record, $params['alias'].'.positioncer') !== 'annule'
 						&& !Hash::get($record, $params['alias'].'.datenotification')
 						&& !Hash::get($record, 'Propodecisioncer66.isvalidcer');
 					break;
-				case 58: 
+				case 58:
 					$access = Hash::get($params, 'haveOrient')
 						&& !Hash::get($params, 'haveOrientEmploi');
 					break;
-				default: 
+				default:
 					$access = true;
 			}
-			
+
 			return $access;
 		}
 
 		/**
 		 * Permission d'accès
-		 * 
+		 *
 		 * @param array $record
 		 * @param array $params
 		 * @return boolean
 		 */
 		protected static function _impression(array $record, array $params) {
 			$params = self::params($params);
-			
+
 			switch ($params['departement']) {
-				case 66: 
+				case 66:
 					$access = Hash::get($record, $params['alias'].'.positioncer') !== 'annule';
 					break;
 				default:
 					$access = true;
 			}
-			
-			return $access;	
+
+			return $access;
 		}
 
 		/**
 		 * Permission d'accès
-		 * 
+		 *
 		 * @param array $record
 		 * @param array $params
 		 * @return boolean
@@ -124,28 +124,28 @@
 
 		/**
 		 * Permission d'accès
-		 * 
+		 *
 		 * @param array $record
 		 * @param array $params
 		 * @return boolean
 		 */
 		protected static function _cancel(array $record, array $params) {
 			$params = self::params($params);
-			
+
 			switch ($params['departement']) {
 				case 66:
 					$access = Hash::get($record, $params['alias'].'.positioncer') !== 'annule';
 					break;
-				default: 
+				default:
 					$access = true;
 			}
-			
+
 			return $access;
 		}
 
 		/**
 		 * Permission d'accès
-		 * 
+		 *
 		 * @param array $record
 		 * @param array $params
 		 * @return boolean
@@ -156,14 +156,14 @@
 
 		/**
 		 * Permission d'accès
-		 * 
+		 *
 		 * @param array $record
 		 * @param array $params
 		 * @return boolean
 		 */
 		protected static function _ficheliaisoncer(array $record, array $params) {
 			$params = self::params($params);
-			
+
 			return Hash::get($record, $params['alias'].'.positioncer') !== 'annule'
 				&& Hash::get($record, 'Propodecisioncer66.isvalidcer')
 			;
@@ -171,14 +171,14 @@
 
 		/**
 		 * Permission d'accès
-		 * 
+		 *
 		 * @param array $record
 		 * @param array $params
 		 * @return boolean
 		 */
 		protected static function _notifbenef(array $record, array $params) {
 			$params = self::params($params);
-			
+
 			return Hash::get($record, $params['alias'].'.positioncer') !== 'annule'
 				&& Hash::get($record, 'Propodecisioncer66.isvalidcer')
 			;
@@ -186,14 +186,14 @@
 
 		/**
 		 * Permission d'accès
-		 * 
+		 *
 		 * @param array $record
 		 * @param array $params
 		 * @return boolean
 		 */
 		protected static function _notificationsop(array $record, array $params) {
 			$params = self::params($params);
-			
+
 			return Hash::get($record, $params['alias'].'.positioncer') !== 'annule'
 				&& Hash::get($record, 'Propodecisioncer66.isvalidcer')
 				&& Hash::get($record, 'Propodecisioncer66.isvalidcer') !== 'N'
@@ -202,14 +202,14 @@
 
 		/**
 		 * Permission d'accès
-		 * 
+		 *
 		 * @param array $record
 		 * @param array $params
 		 * @return boolean
 		 */
 		protected static function _notification(array $record, array $params) {
 			$params = self::params($params);
-			
+
 			return Hash::get($record, $params['alias'].'.positioncer') !== 'annule'
 				&& Hash::get($record, $params['alias'].'.decision_ci') !== 'E'
 			;
@@ -217,7 +217,7 @@
 
 		/**
 		 * Permission d'accès
-		 * 
+		 *
 		 * @param array $record
 		 * @param array $params
 		 * @return boolean
@@ -245,7 +245,7 @@
 
 		/**
 		 * Permission d'accès
-		 * 
+		 *
 		 * @param array $record
 		 * @param array $params
 		 * @return boolean
@@ -258,7 +258,7 @@
 		 * Liste les actions disponnible
 		 * Si une action pointe sur un autre controler, il faut préciser son nom
 		 * ex : Moncontroller.monaction
-		 * 
+		 *
 		 * @param array $params
 		 * @return array
 		 */
@@ -266,14 +266,14 @@
 			$params = self::params($params);
 			$result = self::normalize_actions(
 				array(
-					'add' => array('ajoutPossible' => true), 
-					'view', 
-					'edit', 
-					'impression', 
+					'add' => array('ajoutPossible' => true),
+					'view',
+					'edit',
+					'impression',
 					'filelink'
 				)
 			);
-			
+
 			switch ($params['departement']) {
 				case 66:
 					$result = self::merge_actions($result,
@@ -311,7 +311,7 @@
 					);
 					break;
 			}
-			
+
 			return $result;
 		}
 	}
